requirements.txt
watchdog==4.0.1
supabase==2.5.0
python-dotenv==1.0.1
pdfplumber==0.11.0
